requirements.txt
requests==2.22.0
ndjson==0.3.1
backoff==1.10.0
google-api-core>=1.22.1
pydantic>=1.8,<2.0
shapely
tqdm
geojson
numpy
PILLOW
opencv-python
typeguard
imagesize
pyproj
pygeotile
typing-extensions
pytest-xdist
nbformat~=5.7.0
nbconvert~=7.2.6
azure-storage-blob
loguru